WIP: Install pexpect from PyPI

+++ debian/setup-venv.sh
@@ -3,13 +3,14 @@
 set -euxo pipefail
 
 NAME=$1
-WHEELS_DIR="/builder/securedrop-${NAME}/wheels"
-PIP_ARGS="--ignore-installed --no-index --find-links ${WHEELS_DIR} --no-deps --no-cache-dir --no-build-isolation"
+PIP_ARGS="--ignore-installed --no-binary=:all: --no-deps --no-cache-dir --no-build-isolation"
 
 /usr/bin/python3 -m venv --system-site-packages ./debian/securedrop-${NAME}/opt/venvs/securedrop-${NAME}
-./debian/securedrop-${NAME}/opt/venvs/securedrop-${NAME}/bin/pip install $PIP_ARGS --require-hashes \
-    -r ${NAME}/build-requirements.txt
-./debian/securedrop-${NAME}/opt/venvs/securedrop-${NAME}/bin/pip install $PIP_ARGS ./${NAME}
+if [[ -f "${NAME}/build-requirements.txt" ]]; then
+    ./debian/securedrop-${NAME}/opt/venvs/securedrop-${NAME}/bin/pip install $PIP_ARGS --require-hashes \
+        -r ${NAME}/build-requirements.txt
+fi
+./debian/securedrop-${NAME}/opt/venvs/securedrop-${NAME}/bin/pip install --no-index $PIP_ARGS ./${NAME}
 
 # Adjust paths to reflect installed paths
 find ./debian/securedrop-${NAME}/ -type f -exec sed -i "s#$(pwd)/debian/securedrop-${NAME}##" {} \;
